+++ src/xtd.core/include/xtd/net/sockets/ip_v6_multicast_option.hpp
@@ -43,7 +43,7 @@ namespace xtd {
         /// @brief Initializes a new instance of the xtd::net::sockets::ip_v6_multicast_option class with the specified IP multicast group and the local interface address.
         /// @param group The xtd::net::ip_address of the multicast group.
         /// @param interface_index The address of the interface.
-        ip_v6_multicast_option(const xtd::net::ip_address& group, uint32 interface_index);
+        ip_v6_multicast_option(const xtd::net::ip_address& group, uint64 interface_index);
         /// @}
         
         /// @cond
@@ -65,15 +65,15 @@ namespace xtd {
         ip_v6_multicast_option& group(const xtd::net::ip_address& value) noexcept;
         
         /// @brief Gets the interface index that is associated with a multicast group.
-        /// @return A uint32 value that specifies the address of the interface.
+        /// @return A uint64 value that specifies the address of the interface.
         /// @remarks This property specifies the interface on which data is received or sent.
-        uint32 interface_index() const noexcept;
+        uint64 interface_index() const noexcept;
         /// @brief Sets the interface index that is associated with a multicast group.
-        /// @param value A uint32 value that specifies the address of the interface.
+        /// @param value A uint64 value that specifies the address of the interface.
         /// @return the current instance.
         /// @exception xtd::argument_out_of_range_exception value > 0x00000000FFFFFFFF.
         /// @remarks This property specifies the interface on which data is received or sent.
-        ip_v6_multicast_option& interface_index(uint32 value);
+        ip_v6_multicast_option& interface_index(uint64 value);
         /// @}
         
         /// @name Public Methods
@@ -95,7 +95,7 @@ namespace xtd {
         
       private:
         xtd::net::ip_address group_ = xtd::net::ip_address::ip_v6_none;
-        uint32 interface_index_ = 0;
+        uint64 interface_index_ = 0;
       };
     }
   }
