+++ src/xtd.core/src/xtd/net/sockets/ip_v6_multicast_option.cpp
@@ -8,7 +8,7 @@ using namespace xtd::net::sockets;
 ip_v6_multicast_option::ip_v6_multicast_option(const ip_address& group) : group_ {group} {
 }
 
-ip_v6_multicast_option::ip_v6_multicast_option(const ip_address& group, uint32 interface_index) : group_ {group}, interface_index_ {interface_index} {
+ip_v6_multicast_option::ip_v6_multicast_option(const ip_address& group, uint64 interface_index) : group_ {group}, interface_index_ {interface_index} {
   if (interface_index > 0x00000000FFFFFFFF) throw argument_out_of_range_exception {};
 }
 
@@ -21,11 +21,11 @@ ip_v6_multicast_option& ip_v6_multicast_option::group(const xtd::net::ip_address
   return *this;
 }
 
-uint32 ip_v6_multicast_option::interface_index() const noexcept {
+uint64 ip_v6_multicast_option::interface_index() const noexcept {
   return interface_index_;
 }
 
-ip_v6_multicast_option& ip_v6_multicast_option::interface_index(uint32 value) {
+ip_v6_multicast_option& ip_v6_multicast_option::interface_index(uint64 value) {
   if (value > 0x00000000FFFFFFFF) throw argument_out_of_range_exception {};
   interface_index_ = value;
   return *this;
